Mouse enter & leave of terminal window detected.The root window also calls mouseOut properly on focused widget if this happens.

+++ vterm/terminal.h
@@ -399,7 +399,7 @@ namespace vterm {
          */
         bool scrollable_;
 
-        int historySizeLimit_;
+        size_t historySizeLimit_;
 
         std::deque<std::string> history_;
